add polygon

+++ pojo/request/portal.go
@@ -75,8 +75,8 @@ type TowerStats struct {
 type Sign struct {
 	BaseRequest
 	BasePortalRequest
-	Chain   uint8 `json:"chain" validate:"required"`
-	TokenId int64 `json:"token_id" validate:"required"`
+	Chain   uint64 `json:"chain" validate:"required"`
+	TokenId int64  `json:"token_id" validate:"required"`
 }
 
 type ShelfSign struct {
@@ -110,7 +110,7 @@ type Orders struct {
 	SortTime  uint   `json:"sort_time" validate:"max=2"`
 	Search    string `json:"search" validate:"max=192"`
 
-	ChainId uint8 `json:"chain_id"`
+	ChainId uint64 `json:"chain_id"`
 }
 
 type OrderCancel struct {
@@ -133,7 +133,7 @@ type UserHistory struct {
 	Type              uint8     `json:"type" validate:"required"` //1:transaction history  //2:mint history //3:Listing history
 	FilterTransaction uint8     `json:"filter_transaction"`
 	FilterTime        time.Time `json:"filter_time"`
-	ChainId           uint8     `json:"chain_id"`
+	ChainId           uint64    `json:"chain_id"`
 }
 
 type ExchangePrice struct {
@@ -146,9 +146,9 @@ type ExchangePrice struct {
 type AssetDetail struct {
 	BaseRequest
 	BasePortalRequest
-	AssetId int64 `json:"asset_id"`
-	ChainId uint8 `json:"chain_id"`
-	TokenId int64 `json:"token_id"`
+	AssetId int64  `json:"asset_id"`
+	ChainId uint64 `json:"chain_id"`
+	TokenId int64  `json:"token_id"`
 }
 
 type GameCurrency struct {
@@ -168,14 +168,14 @@ type OrdersGroup struct {
 	SortTime  uint   `json:"sort_time" validate:"max=2"`
 	Search    string `json:"search" validate:"max=192"`
 
-	ChainId uint8 `json:"chain_id"`
+	ChainId uint64 `json:"chain_id"`
 }
 
 type OrdersGroupDetail struct {
 	BaseRequest
 	GroupName string `json:"group_name" validate:"required"`
 	SortPrice uint   `json:"sort_price" validate:"max=2"`
-	ChainId   uint8  `json:"chain_id"`
+	ChainId   uint64 `json:"chain_id"`
 }
 
 type SendCode struct {
@@ -186,7 +186,7 @@ type SendCode struct {
 type PaperMint struct {
 	BaseRequest
 	TokenId       int64  `json:"token_id" validate:"required"`
-	ChainId       uint8  `json:"chain_id" validate:"required"`
+	ChainId       uint64 `json:"chain_id" validate:"required"`
 	Email         string `json:"email"  validate:"required"`
 	WalletAddress string `json:"wallet_address"  validate:"required"`
 }
@@ -230,5 +230,5 @@ type OrdersOfficial struct {
 	SortTime  uint   `json:"sort_time" validate:"max=2"`
 	Search    string `json:"search" validate:"max=192"`
 
-	ChainId uint8 `json:"chain_id" validate:"required"`
+	ChainId uint64 `json:"chain_id" validate:"required"`
 }
